feat(select): separate file for select stories
